pyproject.toml
[project]
name = "hdemucs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flax==0.10.4",
    "jupyter>=1.1.1",
    "orbax-checkpoint==0.11.9",
    "pytest>=8.3.4",
    "soundfile>=0.13.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.48.3",
    "treescope>=0.1.9",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.1",
    "pytest-benchmark>=5.1.0",
    "ruff>=0.11.2",
]
